+++ notify.h
@@ -1,0 +1,7 @@
+#ifndef NOTIFY_H
+#define NOTIFY_H
+#include <stdbool.h>
+#include <stdio.h>
+int notify(char* head, char* message);
+void ghost(bool debug);
+#endif
